% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_data.R
\name{annotate_data}
\alias{annotate_data}
\title{annotate_data}
\usage{
annotate_data(data, file_root, fields = c("pep_id", "pos_start",
  "pos_end", "UniProt_acc", "pep_aa", "taxon_species", "gene_symbol",
  "product"), md_path)
}
\arguments{
\item{data}{Un-annotated data.}

\item{file_root}{Filename to write.}

\item{fields}{What information to populate from annotation file.}

\item{md_path}{Metadata directory.}
}
\description{
Add peptide/protein annotation to data file.
}
