% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_order.R
\name{check_order}
\alias{check_order}
\title{Maintains specified order of panlibrary data after a call to dplyr::bind_rows.}
\usage{
check_order(libs, data)
}
\arguments{
\item{libs}{Character vector of library ID's in order}

\item{data}{Merged data to re-sort.}
}
\description{
Maintains specified order of panlibrary data after a call to dplyr::bind_rows.
}
