% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_polycl.R
\name{compute_polycl}
\alias{compute_polycl}
\title{compute_polycl}
\usage{
compute_polycl(data, annot, pairs, verbose = TRUE)
}
\arguments{
\item{data}{List with first element a character vector of library names and
subsequent elements data frames of data for corresponding libraries.}

\item{annot}{Phiplist of annotation files.}

\item{pairs}{Phiplist of blast alignment pair files.}

\item{verbose}{Logical whether or not to print progress.}
}
\description{
For multilibrary hits data, identify libraries, and call library_polycl
for each.
}
