% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_promax.R
\name{compute_promax}
\alias{compute_promax}
\title{compute_promax}
\usage{
compute_promax(data, annot)
}
\arguments{
\item{data}{List with first element a character vector of library names and
subsequent elements data frames of data for corresponding libraries.}

\item{annot}{Phiplist of annotation files.}
}
\description{
For multilibrary hits data, identify libraries, and call library_promax
for each.
}
