% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prosum.R
\name{compute_prosum}
\alias{compute_prosum}
\title{compute_prosum}
\usage{
compute_prosum(data, file_root, md_path)
}
\arguments{
\item{data}{List with first element a character vector of library names and
subsequent elements data frames of data for corresponding libraries.}

\item{file_root}{Filename to write.}

\item{md_path}{Metadata directory.}
}
\description{
For multilibrary data, identify libraries, and call library_prosum
for each.
}
