% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_plan.R
\name{define_plan}
\alias{define_plan}
\title{define_plan}
\usage{
define_plan(md_path, pairs_path)
}
\arguments{
\item{md_path}{Directory path to metadata peptide annotation files. Not implemented. need Drake wildcards.}

\item{pairs_path}{Directory path to intraprotein BLAST pairs files. Not implemented. need Drake wildcards.}
}
\description{
Initialize and return drake_plan object
}
