% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emphasize_hit_data.R
\name{emphasize_hit_data}
\alias{emphasize_hit_data}
\title{emphasize_hit_data}
\usage{
emphasize_hit_data(data, hits, default)
}
\arguments{
\item{data}{Data frame of raw data to use.}

\item{hits}{Data frame of binary data with 1 for hits and 0 for non-hits.}

\item{default}{Numeric value to replace non-hit values with.}
}
\description{
Set non-hit values to a default value.
}
