% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_samples.R
\name{gather_samples}
\alias{gather_samples}
\title{read in multiple input files and extract specified columns}
\usage{
gather_samples(sources, col.names, retain.first.col = TRUE)
}
\arguments{
\item{sources}{Character vector of filepaths to read in. All source files should contain the same type of data - enrichments, normalized read counts, etc.}

\item{col.names}{Character vector of columns to extract. Vector order should match desired order in output.}

\item{retain.first.col}{Logical whether to also include the furst column of the first specified source file. Often this column will contain peptide names of interest.}
}
\description{
read in multiple input files and extract specified columns
}
