% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getparam.R
\name{getparam}
\alias{getparam}
\title{Function to retreive value of parameter from param name.}
\usage{
getparam(params, param)
}
\arguments{
\item{params}{Data frame of parameters with two columns param and value.}

\item{param}{Character string corresponding to a value in param$param.}
}
\description{
Function to retreive value of parameter from param name.
}
