% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyclonal.R
\name{independence_filter}
\alias{independence_filter}
\title{Compute independent hits usingg network graphing.}
\usage{
independence_filter(edge, vertex)
}
\arguments{
\item{edge}{Alignment dictionary}

\item{vertex}{List of nodes to compute on.}
}
\description{
Compute independent hits usingg network graphing.
}
