% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyclonal.R
\name{library_polycl}
\alias{library_polycl}
\title{For one peptide library, identify proteins from annotation file, and then
call protein_polycl for each protein.}
\usage{
library_polycl(data, annot, pairs, method = "independence_filter",
  verbose = TRUE)
}
\arguments{
\item{data}{Data frame with one library's hits data}

\item{annot}{Annotation file.}

\item{pairs}{Data frame of BLAST aligning pairs within each protein.}

\item{method}{Character string signifying which scoring method to use for evaluation. independence_filter or old_method.}

\item{verbose}{Logical whether or not to print progress.}
}
\description{
For one peptide library, identify proteins from annotation file, and then
call protein_polycl for each protein.
}
