% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_promax.R
\name{library_promax}
\alias{library_promax}
\title{library_promax}
\usage{
library_promax(data, annot)
}
\arguments{
\item{data}{Data frame with one library's data.}

\item{annot}{Annotation file.}
}
\description{
For one peptide library, identify proteins from annotation file, and then
call protein_promax for each protein.
}
