% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_annotations.R
\name{merge_annotations}
\alias{merge_annotations}
\title{Merge epitopefindr fields from library annotation files}
\usage{
merge_annotations(enrichment_annotations)
}
\arguments{
\item{enrichment_annotations}{Annotation files}
}
\description{
Merge epitopefindr fields from library annotation files
}
