% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbar.R
\name{pbar}
\alias{pbar}
\title{pbar}
\usage{
pbar(min, max)
}
\arguments{
\item{min}{Progress bar minimum value.}

\item{max}{Progress bar maximum value.}
}
\description{
Wrapper for txtProgressBar with some defualt settings.
}
