% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phipmake.R
\docType{package}
\name{phipmake}
\alias{phipmake}
\title{phipmake: PhIP-seq Data Analysis Pipeline.}
\description{
PhIP-seq is an antibody profiling assay.
Phipmake uses ropensci's drake package.
}
\details{
To keep track of multilibrary data, phipmake uses specially formatted lists
that we are calling phiplists.
}
