% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_avarda_names.R
\name{prepare_avarda_names}
\alias{prepare_avarda_names}
\title{swap out u_pep_id names for older pep_id names to improve AVARDA speed}
\usage{
prepare_avarda_names(
  data,
  annotations = NULL,
  metadata_path = NULL,
  replacement_column = "pep_id"
)
}
\arguments{
\item{data}{Phiplist of data to modify.}

\item{annotations}{Phiplist of annotation files. Don't use with metadata_path param.}

\item{metadata_path}{Metadata directory. Don't use with annotations param.}

\item{replacement_column}{Column of metadata with which to replace current first column of data.}
}
\description{
swap out u_pep_id names for older pep_id names to improve AVARDA speed
}
