% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_blast_pairs.R
\name{prepare_blast_pairs}
\alias{prepare_blast_pairs}
\title{Convert BLAST+ alignment table to two outputs. First, a full symmetrical two
column table of aligning peptides. Second, a subset of the first, including
only intra-protein alignments, for which protein identity is specified by
an annotation file.}
\usage{
prepare_blast_pairs(blast.file, annot, name1, name2)
}
\arguments{
\item{blast.file}{directory path to blast file e.g. 1/PhageomeLa_000_BLAST.txt}

\item{annot}{Annotation file.}

\item{name1}{file path & name for all-alignment output file tsv}

\item{name2}{file path & name for intraprotein-alignment output file tsv}
}
\description{
Convert BLAST+ alignment table to two outputs. First, a full symmetrical two
column table of aligning peptides. Second, a subset of the first, including
only intra-protein alignments, for which protein identity is specified by
an annotation file.
}
