% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_polycl.R
\name{protein_polycl}
\alias{protein_polycl}
\title{protein_polycl}
\usage{
protein_polycl(data, pairs, margin = 2)
}
\arguments{
\item{data}{Data frame of numeric values for which to calculate maximum.}

\item{pairs}{2 column table of aligning pairs that should not be double-counted for polyclonal score.}

\item{margin}{Numeric 1 for row 2 for column.}
}
\description{
Take data frame input and calculate the polyclonal score in each column or row.
}
