% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_promax.R
\name{protein_promax}
\alias{protein_promax}
\title{Take data frame input and calculate the maximum value in each column or row.}
\usage{
protein_promax(data, margin = 2)
}
\arguments{
\item{data}{Data frame of numeric values for which to calculate maximum.}

\item{margin}{Numeric 1 for row 2 for column.}
}
\description{
Take data frame input and calculate the maximum value in each column or row.
}
