% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosum.R
\name{protein_prosum}
\alias{protein_prosum}
\title{Take data frame input and calculate the sum of values in each column or row.}
\usage{
protein_prosum(data, margin = 2)
}
\arguments{
\item{data}{Data frame of numeric values for which to calculate sum}

\item{margin}{Numeric 1 for row 2 for column.}
}
\description{
Take data frame input and calculate the sum of values in each column or row.
}
