% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_annot.R
\name{read_annot}
\alias{read_annot}
\title{Read annotation file based on md path and library name.}
\usage{
read_annot(libname, metadata_path)
}
\arguments{
\item{libname}{Library name.}

\item{metadata_path}{Metadata directory.}
}
\description{
Read annotation file based on md path and library name.
}
