% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_annot_list.R
\name{read_annot_list}
\alias{read_annot_list}
\title{read_annot_list}
\usage{
read_annot_list(libs, metadata_path)
}
\arguments{
\item{libs}{Character vector of library names.}

\item{md_path}{Metadata directory as character string.}
}
\value{
phiplist of annotations. First element is character vector of
libraries from libs input. Remaining elements are corresponding annotation
files.
}
\description{
Read annotation files based on library names and metadata path.
}
