% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pairs.R
\name{read_pairs}
\alias{read_pairs}
\title{read_pairs}
\usage{
read_pairs(libname, pairs_path)
}
\arguments{
\item{libname}{Library name.}

\item{pairs_path}{Pairs data directory.}
}
\description{
Read BLAST pairs file based on md path and library name.
}
