% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pairs.R
\name{read_pairs_list}
\alias{read_pairs_list}
\title{Read blast alignment pair files based on library names and metadata path.}
\usage{
read_pairs_list(libs, pairs_path)
}
\arguments{
\item{libs}{Character vector of library names.}

\item{pairs_path}{Metadata directory as character string.}
}
\value{
phiplist of blast pairs First element is character vector of
libraries from libs input. Remaining elements are corresponding alignment
files.
}
\description{
Read blast alignment pair files based on library names and metadata path.
}
