% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_epitopefindr.R
\name{run_epitopefindr}
\alias{run_epitopefindr}
\title{Assemble hit files and call epitopefindr}
\usage{
run_epitopefindr(
  hits_foldchange,
  annotation_merged_df,
  epitopefindr_latex,
  parallel = FALSE,
  ...
)
}
\description{
Assemble hit files and call epitopefindr
}
