% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{split_data}
\usage{
split_data(data)
}
\arguments{
\item{data}{Multilibrary data frame.}
}
\value{
List with n+1 elements for a data file with n sublibraries. The first
element is a character vector of sublibrary names in order. The subsequent
elements are data frames corresponding to the specified libraries, in
corresponding order.
}
\description{
Separate multilibrary peptide data into a phipmake list.
}
