% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_data.R
\name{subset_data}
\alias{subset_data}
\title{subset_data}
\usage{
subset_data(data, sub, filename)
}
\arguments{
\item{data}{Multilibrary data frame.}

\item{sub}{Subset term to grep for in first column of data.}

\item{filename}{Filename to write.}
}
\description{
Pull out specific sublibrary data from a larger data frame based on first column grep.
}
