% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimDictionary.R
\name{trimDictionary}
\alias{trimDictionary}
\title{trim alignment dictionary to not include self alignments or redundant values}
\usage{
trimDictionary(edge)
}
\arguments{
\item{edge}{Input dictionary.}
}
\description{
trim alignment dictionary to not include self alignments or redundant values
}
