% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_data}
\alias{write_data}
\title{A wrapper for data.table::fwrite to write file to drive.}
\usage{
write_data(data, filename)
}
\arguments{
\item{data}{Data to write. If class of data is data.frame or data.table, then
write_data will fwrite. If class of data is a list, write_data will assume
that data is a phiplist and write the second through final elements as
separate files.}

\item{filename}{Filename or filenames to write.}
}
\description{
A wrapper for data.table::fwrite to write file to drive.
}
