% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_drake_params.R
\name{write_drake_params}
\alias{write_drake_params}
\title{write_drake_params}
\usage{
write_drake_params(dir, screen_name, counts_filename = "counts.csv",
  counts_type = "Counts", foldchange_filename = "fold_change.csv",
  foldchange_type = "FoldChange",
  enrichment_filename = "enrichment.csv",
  enrichment_type = "Enrichment", hits_filename = "Hits.csv",
  metadata_path = "/data/hlarman1/PhIPdb/Metadata/PeptideLibraries",
  output_extension = "tsv")
}
\arguments{
\item{dir}{PhIP-seq project directory.}

\item{screen_name}{Identifier for multiplexed screen, e.g. phipseq_0001.}

\item{counts_filename}{Name of demultiplexed, merged reads file in screen directory.}

\item{counts_type}{Identifier for type of counts data to append in output file names.}

\item{foldchange_filename}{Name of demultiplexed, merged foldchange file in screen directory.}

\item{foldchange_type}{Identifier for type of counts data to append in output file names.}

\item{enrichment_filename}{Name of demultiplexed, merged enrichment file in screen directory.}

\item{enrichment_type}{Identifier for type of enrichment data to append in output file names.}

\item{hits_filename}{Name of demultiplexed, merged hits file in screen directory.}

\item{metadata_path}{Directory path string to metadata folder containing peptide files and intrapeptide blast alignment files.}

\item{output_extension}{Extension for tab-separated output files.}
}
\description{
write default values to drake_params.tsv for phipmake pipeline
}
