% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_data.R
\name{annotate_data}
\alias{annotate_data}
\title{Add peptide/protein annotation to data file.}
\usage{
annotate_data(data, annotations = NULL, metadata_path = NULL,
  fields = c("pep_id", "pos_start", "pos_end", "UniProt_acc", "pep_aa",
  "taxon_genus", "taxon_species", "gene_symbol", "product"),
  pep_fields = c("pep_id", "pos_start", "pos_end", "pep_aa"),
  parallel = FALSE)
}
\arguments{
\item{data}{Phiplist of unannotated edata.}

\item{annotations}{Phiplist of annotation files. Don't use with metadata_path param.}

\item{metadata_path}{Metadata directory. Don't use with annotations param.}

\item{fields}{What information to populate from annotation file.}

\item{parallel}{Logical whether or not to use future and future.apply to parallelize individual library computations.}

\item{libname}{Library name contained in data.}
}
\value{

}
\description{
Add peptide/protein annotation to data file.
}
