% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_polycl.R
\name{compute_polycl}
\alias{compute_polycl}
\title{compute_polycl}
\usage{
compute_polycl(data, file_root, md_path, pairs_path)
}
\arguments{
\item{data}{List with first element a character vector of library names and
subsequent elements data frames of data for corresponding libraries.}

\item{file_root}{Filename to write.}

\item{md_path}{Metadata directory.}

\item{md_path}{BLAST pairs directory.}
}
\description{
For multilibrary hits data, identify libraries, and call library_polycl
for each.
}
