% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_polycl.R
\name{compute_polycl}
\alias{compute_polycl}
\title{For multilibrary hits data, identify libraries, and call library_polycl
for each.}
\usage{
compute_polycl(data, annot, pairs, method = "independence_filter",
  verbose = TRUE, parallel = FALSE)
}
\arguments{
\item{data}{List with first element a character vector of library names and
subsequent elements data frames of data for corresponding libraries.}

\item{annot}{Phiplist of annotation files.}

\item{pairs}{Phiplist of blast alignment pair files.}

\item{method}{Character string signifying which scoring method to use for evaluation. independence_filter or old_method.}

\item{verbose}{Logical whether or not to print progress.}

\item{parallel}{Logical whether or not to use foreach and doParallel to parallelize individual library computations.}
}
\description{
For multilibrary hits data, identify libraries, and call library_polycl
for each.
}
