% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promax.R
\name{compute_promax}
\alias{compute_promax}
\title{For multilibrary hits data, identify libraries, and call library_promax
for each.}
\usage{
compute_promax(data, annot, verbose = TRUE, parallel = FALSE)
}
\arguments{
\item{data}{List with first element a character vector of library names and
subsequent elements data frames of data for corresponding libraries.}

\item{annot}{Phiplist of annotation files.}

\item{verbose}{Logical whether or not to print progress.}

\item{parallel}{Logical whether or not to use foreach and doParallel to parallelize individual library computations.}
}
\description{
For multilibrary hits data, identify libraries, and call library_promax
for each.
}
