% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosum.R
\name{compute_prosum}
\alias{compute_prosum}
\title{For multilibrary data, identify libraries, and call library_prosum
for each.}
\usage{
compute_prosum(data, annot, verbose = TRUE, parallel = FALSE)
}
\arguments{
\item{data}{List with first element a character vector of library names and
subsequent elements data frames of data for corresponding libraries.}

\item{annot}{Phiplist of annotation files.}

\item{verbose}{Logical whether or not to print progress.}

\item{parallel}{Logical whether or not to use future and future.apply to parallelize individual library computations.}
}
\description{
For multilibrary data, identify libraries, and call library_prosum
for each.
}
