% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_plan.R
\name{define_plan}
\alias{define_plan}
\title{Initialize and return drake_plan object. New for github drake 7.0+.}
\usage{
define_plan(
  params_path = "drake_params.tsv",
  runCounts = TRUE,
  runPairwise = FALSE,
  runFoldChange = TRUE,
  runEnrichment = TRUE,
  runHits = TRUE,
  runPolyclonal = TRUE,
  runAVARDA = FALSE,
  runEpitopefindr = FALSE,
  epitopefindr_latex = FALSE
)
}
\arguments{
\item{params_path}{Path to project parameters file that can be created using the write_drake_params function.}

\item{runCounts}{Logical controlling whether or not to define counts targets.}

\item{runPairwise}{Logical controlling whether or not to define pairwise targets.}

\item{runFoldChange}{Logical controlling whether or not to define EdgeR fold change targets.}

\item{runEnrichment}{Logical controlling whether or not to define enrichment targets.}

\item{runPolyclonal}{Logical controlling whether or not to define enrichment targets.}

\item{runAVARDA}{Logical controlling whether or not to define AVARDA targets.}

\item{runEpitopefindr}{Logical controlling whether or not to run epitopefindr analysis on hits.}

\item{epitopefindr_latex}{Logical controlling whether to generate LaTeX from epitopefindr multiple sequence alignments.}
}
\description{
Initialize and return drake_plan object. New for github drake 7.0+.
}
