% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_plan.R
\name{define_plan}
\alias{define_plan}
\title{Initialize and return drake_plan object. New for github drake 7.0+.}
\usage{
define_plan(params_path = "drake_params.tsv")
}
\arguments{
\item{params_path}{Path to project parameters file that can be created using the write_drake_params function.}
}
\description{
Initialize and return drake_plan object. New for github drake 7.0+.
}
