% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_polycl.R
\name{library_polycl}
\alias{library_polycl}
\title{library_polycl}
\usage{
library_polycl(data, annot, pairs, verbose = TRUE)
}
\arguments{
\item{data}{Data frame with one library's hits data}

\item{annot}{Annotation file.}

\item{pairs}{Data frame of BLAST aligning pairs within each protein.}

\item{verbose}{Logical whether or not to print progress.}
}
\description{
For one peptide library, identify proteins from annotation file, and then
call protein_polycl for each protein.
}
