% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_prosum.R
\name{library_prosum}
\alias{library_prosum}
\title{library_prosum}
\usage{
library_prosum(data, annot)
}
\arguments{
\item{data}{Data frame with one library's data.}

\item{annot}{Annotation file.}
}
\description{
For one peptide library, identify proteins from annotation file, and then
call protein_prosum for each protein.
}
