% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{split_data}
\usage{
split_data(data, file_root)
}
\arguments{
\item{data}{Multilibrary data frame.}

\item{file_root}{Filename to write.}
}
\description{
Separate multilibrary peptide data into a a list of files for individual libraries.
}
